--- Homework/TagsCloud.Visualization/ImagesSaver/IImageSaver.cs
@@ -1,9 +1,0 @@
-﻿using System.Drawing;
-
-namespace TagsCloud.Visualization.ImagesSaver
-{
-    public interface IImageSaver
-    {
-        void Save(Image image);
-    }
-}
