+++ Homework/TagsCloud.Visualization/ImagesSavior/IImageSavior.cs
@@ -1,0 +1,9 @@
+﻿using System.Drawing;
+
+namespace TagsCloud.Visualization.ImagesSavior
+{
+    public interface IImageSavior
+    {
+        void Save(Image image);
+    }
+}
